Replaced RecipeChoice class with interface+record
